requirement.txt
asgiref==3.5.2
cssselect2==0.6.0
dj-database-url==1.0.0
Django==3.1.14
django-classy-tags==3.0.1
django-cms==3.8.0
django-filer==2.2.3
django-formtools==2.3
django-js-asset==2.0.0
django-mptt==0.13.4
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.13.1
easy-thumbnails==2.8.3
html5lib==1.1
lxml==4.9.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
tzdata==2022.2
tzlocal==4.2
Unidecode==1.1.2
webencodings==0.5.1
